requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
certifi==2017.1.23
click==6.7
Flask==0.12
future==0.16.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
pbr==2.0.0
PyMySQL==0.7.10
pyparsing==2.2.0
python-telegram-bot==5.3.0
requests==2.13.0
six==1.10.0
stevedore==1.21.0
urllib3==1.20
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
Werkzeug==0.12.1
